Add initial support for YAML and TOML
There are a few TODOs including the need for YAML and TOML IDs. There
needs to be some testing to ensure that the method implemented here
is reliable as well.

A few rudimentary changes have been made to try and improve performance
just in-case there are any problems. Additionally, YAML processing is
strictly limited to files with an explicit ('---') header as otherwise
the YAML handler will pretty much try and process any string value.

Tests have been updated to match the new control logic.