Make analysis more robust
In cases where data is non-complex, i.e. primitive data types the
analysis does not have to look as deeply as the data as we've anticipated.
That being said, neither scenario should result in an exception. This commit
handles more simple cases better and provides a better harnes for testing
exceptions in the future.