Provide test exceptions for Windows
I cannot find the incantation needed to get libmagic working on Windows.
As such, it's disabled here and the appropriate pathways are created in
the GitHub action for the unit tests that lets tests run correctly for
the Windows platform. Compressed JSONL will not be identified on Windows
at present but we should be able to make this work with FIDO or some
other lightweight format identification library at some point in the future.